+++ src/core/CookieManager.h
@@ -6,6 +6,8 @@
 #include <QObject>
 #include <QString>
 
+#include <atomic>
+
 class QWebEngineCookieStore;
 
 class CookieManager : public QObject
@@ -14,6 +16,7 @@ class CookieManager : public QObject
 
 public:
     explicit CookieManager(QWebEngineCookieStore *store, QObject *parent = nullptr);
+    void setBlockThirdPartyCookies(bool enabled);
 
     Q_INVOKABLE QString getCookiesJson() const;
     Q_INVOKABLE void deleteByIndex(int index);
@@ -29,7 +32,10 @@ private slots:
     void onCookieRemoved(const QNetworkCookie &cookie);
 
 private:
+    void finishReload(int remainingRetries);
+
     QWebEngineCookieStore *m_store = nullptr;
     QList<QNetworkCookie> m_cookies;
     bool m_loading = false;
+    std::atomic_bool m_blockThirdPartyCookies = false;
 };